test(treesitter): split parser and query tests
